Separate screen logic into its own class
