+++ src/Menu.php
@@ -7,6 +7,8 @@
 
 namespace Automattic\WooCommerce\Navigation;
 
+use Automattic\WooCommerce\Navigation\Screen;
+
 /**
  * Contains logic for the WooCommerce Navigation menu.
  */
@@ -61,58 +63,11 @@ class Menu {
 	protected static $menu_items = array();
 
 	/**
-	 * Screen IDs of registered pages.
-	 *
-	 * @var array
-	 */
-	protected static $screen_ids = array();
-
-	/**
-	 * Registered post types.
-	 *
-	 * @var array
-	 */
-	protected static $post_types = array();
-
-	/**
 	 * Registered callbacks or URLs with migration boolean as key value pairs.
 	 *
 	 * @var array
 	 */
 	protected static $callbacks = array();
-
-	/**
-	 * Check if we're on a WooCommerce page
-	 *
-	 * @return bool
-	 */
-	public static function is_woocommerce_page() {
-		global $pagenow, $plugin_page;
-
-		// Get post type if on a post screen.
-		$post_type = '';
-		if ( in_array( $pagenow, array( 'edit.php', 'post.php', 'post-new.php' ), true ) ) {
-			if ( isset( $_GET['post'] ) ) { // phpcs:ignore CSRF ok.
-				$post_type = get_post_type( (int) $_GET['post'] ); // phpcs:ignore CSRF ok.
-			} elseif ( isset( $_GET['post_type'] ) ) { // phpcs:ignore CSRF ok.
-				$post_type = sanitize_text_field( wp_unslash( $_GET['post_type'] ) ); // phpcs:ignore CSRF ok.
-			}
-		}
-		$post_types = apply_filters( 'woocommerce_navigation_post_types', self::$post_types );
-
-		// Get current screen ID.
-		$current_screen = get_current_screen();
-		$screen_ids     = apply_filters( 'woocommerce_navigation_screen_ids', self::$screen_ids );
-
-		if (
-			in_array( $post_type, $post_types, true ) ||
-			in_array( $current_screen->id, self::$screen_ids, true )
-		) {
-			return true;
-		}
-
-		return false;
-	}
 
 	/**
 	 * Get class instance.
@@ -130,21 +85,6 @@ class Menu {
 	public function init() {
 		add_action( 'admin_menu', array( $this, 'add_settings' ), 20 );
 		add_filter( 'add_menu_classes', array( $this, 'migrate_menu_items' ) );
-		add_filter( 'admin_body_class', array( $this, 'add_body_class' ) );
-	}
-
-	/**
-	 * Add navigation classes to body.
-	 *
-	 * @param string $classes Classes.
-	 * @return string
-	 */
-	public function add_body_class( $classes ) {
-		if ( self::is_woocommerce_page() ) {
-			$classes .= ' has-woocommerce-navigation';
-		}
-
-		return $classes;
 	}
 
 	/**
@@ -160,6 +100,31 @@ class Menu {
 			return $callback;
 		}
 		return 'admin.php?page=' . $callback;
+	}
+
+	/**
+	 * Get the parent key if one exists.
+	 *
+	 * @param string $url Callback or URL.
+	 * @return string|null
+	 */
+	public static function get_parent_key( $callback ) {
+		global $submenu;
+
+		// This is already a parent item.
+		if ( isset( $submenu[ $callback ] ) ) {
+			return null;
+		}
+
+		foreach ( $submenu as $key => $menu ) {
+			foreach ( $menu as $item ) {
+				if ( $item[ self::CALLBACK ] === $callback ) {
+					return $key;
+				}
+			}
+		}
+
+		return null;
 	}
 
 	/**
@@ -214,31 +179,6 @@ class Menu {
 	}
 
 	/**
-	 * Get the parent menu item if one exists.
-	 *
-	 * @param string $url URL or callback.
-	 * @return string|null
-	 */
-	public static function get_parent_item( $url ) {
-		global $submenu;
-
-		// This is already a parent item.
-		if ( isset( $submenu[ $url ] ) ) {
-			return null;
-		}
-
-		foreach ( $submenu as $key => $menu ) {
-			foreach ( $menu as $item ) {
-				if ( $item[ self::CALLBACK ] === $url ) {
-					return $key;
-				}
-			}
-		}
-
-		return null;
-	}
-
-	/**
 	 * Hides all WP admin menus items and adds screen IDs to check for new items.
 	 *
 	 * @param array $menu Menu items.
@@ -279,25 +219,10 @@ class Menu {
 		}
 
 		foreach ( array_keys( self::$callbacks ) as $callback ) {
-			self::add_screen_id( $callback );
+			Screen::add_screen( $callback );
 		}
 
 		return $menu;
-	}
-
-	/**
-	 * Adds a screen ID to the list and automatically finds the parent if none is given.
-	 *
-	 * @param string      $url URL or callback for page.
-	 * @param string|null $parent Parent slug.
-	 */
-	public static function add_screen_id( $url, $parent = null ) {
-		global $submenu;
-
-		if ( ! $parent ) {
-			$parent = self::get_parent_item( $url );
-		}
-		self::$screen_ids[] = get_plugin_page_hookname( $url, $parent );
 	}
 
 	/**
